requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.5.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.12
Flask-WhooshAlchemy==0.56
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
SQLAlchemy==1.0.8
Tempita==0.5.2
WTForms==2.0.2
Werkzeug==0.10.4
Whoosh==2.7.0
alembic==0.8.2
argparse==1.3.0
blinker==1.4
decorator==4.0.2
gunicorn==19.3.0
itsdangerous==0.24
pbr==1.6.0
psycopg2==2.6.1
python-editor==0.4
python-openid==2.2.5
six==1.9.0
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
stevedore==1.7.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.0
wsgiref==0.1.2
